Rename: QueryProvider 파일 위치 이동
